fix settings and match screen layouts
